- Updated code to match ElFinder.net 0.9.5 version http://elfinder.codeplex.com/releases/view/96866
- Added the ability to Lock specific folders using the LockedFolders property.
https://elfinder.codeplex.com/discussions/442569

+++ elFinder.Net/DownloadFileResult.cs
@@ -17,38 +17,48 @@ namespace ElFinder
         public override void ExecuteResult(ControllerContext context)
         {
             HttpResponseBase response = context.HttpContext.Response;
-            string fileName;
-            string fileNameEncoded = HttpUtility.UrlEncode(File.Name);
+            HttpRequestBase request = context.HttpContext.Request;
+            if (!HttpCacheHelper.IsFileFromCache(File, request, response))
+            {
 
-            if (context.HttpContext.Request.UserAgent.Contains("MSIE")) // IE < 9 do not support RFC 6266 (RFC 2231/RFC 5987)
-            {
-                fileName = "filename=\"" + fileNameEncoded + "\"";
+                string fileName;
+                string fileNameEncoded = HttpUtility.UrlEncode(File.Name);
+
+                if (context.HttpContext.Request.UserAgent.Contains("MSIE")) // IE < 9 do not support RFC 6266 (RFC 2231/RFC 5987)
+                {
+                    fileName = "filename=\"" + fileNameEncoded + "\"";
+                }
+                else
+                {
+                    fileName = "filename*=UTF-8\'\'" + fileNameEncoded; // RFC 6266 (RFC 2231/RFC 5987)
+                }
+                string mime;
+                string disposition;
+                if (IsDownload)
+                {
+                    mime = "application/octet-stream";
+                    disposition = "attachment; " + fileName;
+                }
+                else
+                {
+                    mime = Helper.GetMimeType(File);
+                    disposition = (mime.Contains("image") || mime.Contains("text") || mime == "application/x-shockwave-flash" ? "inline; " : "attachment; ") + fileName;
+                }
+
+                response.ContentType = mime;
+                response.AppendHeader("Content-Disposition", disposition);
+                response.AppendHeader("Content-Location", File.Name);
+                response.AppendHeader("Content-Transfer-Encoding", "binary");
+                response.AppendHeader("Content-Length", File.Length.ToString());
+                response.WriteFile(File.FullName);
+                response.End();
+                response.Flush();
             }
             else
             {
-                fileName = "filename*=UTF-8\'\'" + fileNameEncoded; // RFC 6266 (RFC 2231/RFC 5987)
+                response.ContentType = IsDownload ? "application/octet-stream" : Helper.GetMimeType(File);
+                response.End();
             }
-            string mime;
-            string disposition;
-            if (IsDownload)
-            {
-                mime = "application/octet-stream";
-                disposition = "attachment; " + fileName;
-            }
-            else
-            {
-                mime = Helper.GetMimeType(File);
-                disposition = mime.Contains("image") || mime.Contains("text") || mime == "application/x-shockwave-flash" ? "inline" : "attachment; " + fileName;
-            }           
-			
-            response.ContentType = mime;
-            response.AppendHeader("Content-Disposition", disposition);
-            response.AppendHeader("Content-Location", File.Name);
-            response.AppendHeader("Content-Transfer-Encoding", "binary");
-            response.AppendHeader("Content-Length", File.Length.ToString());
-            response.WriteFile(File.FullName);
-            response.End();
-            response.Flush();
         }
     }
 }
